Expand `DistributedPerformanceTest` configuration
This commit adds the ability to create a distributed performance test which overrides the baselines,
warmups, runs and checks. It requires a change to the dev infrastructure to work as expected, or the
new parameters are going to be ignored.

This is a step towards a new task defining an historical build.