Ignore errors when measurements checks are set to `none`
This commit lets us ignore build failures when the performance checks mode is set to `none`. This means
that if a template is wrong, or a particular version of Gradle is not compatible with a template, and
that an iteration of a performance test fails, measurement is ignored *and* the performance test will
not fail.

This is going to be useful for historical builds, where we want to override the baselines for all tests.