Fix native binaries documentation around platform variant axis selection
Starting with Gradle 2.3, binaries for all declared platforms are not
built anymore. Only the 'current' platform is used. One have to
explicitly declare the `targetPlatform`s on components.

This commit fixes one paragraph in the Native Binaries chapter of the
user guide that still contained an affirmation describing the old
behaviour.