Convert AsyncDispatch test from JMock to Groovy
Converted MultithreadedTestCase to a Rule in order
to use this for this test. It looks like we have multiple
classes for testing concurrent code. There is

 - org.gradle.test.fixtures.concurrent.ConcurrentSpec (67 usages)
 - org.gradle.test.fixtures.ConcurrentTestUtil (~60 usages)
 - org.gradle.util.ConcurrentSpecification (10 subclasses), which uses ConcurrentTestUtil

It would be nice to have one class going forward. MultithreadedTestCase has currently
four subclasses, of which two are JMock tests. It would be nice to get rid of this one at
least, but to me it is not obvious how the feature waiting on a `tick` is covered by
the other test utilities.