Make sure the performance report is always cleaned
This commit makes sure that we run `cleanPerformanceReport` as soon as we ask for a new performance test. This will fix the issue of
"stale" reports seen on CI when a test task fails: the report of the previous execution was shown. The issue is that the various
`cleanXXX` tasks are no real tasks, but rules, so it's not possible to tell that the `cleanXXXPerfTest` task depends on `cleanPerformanceReport`.
Instead, we're declaring that each performance test task depends on both the `clean` and report task.