Ignore environment as input for Exec task
On Mac OS there are some environment variables which
change from build to build. This would mean that the
Exec task would never be up to date on Mac OS. Given
that even on Unix environment variables would change
we need some deeper modeling of this fact - only
adding `@Input` to `getEnvironment()` does not
model the inputs correctly.

The existing tests do not capture that problem because
we do not run them on Mac OS in CI. The problems only
are visible when the tests are run in `forking` mode.

To restore the old behaviour `@Input` has been removed
on `getEnvironment()`.

GRADLE-3528 #resolve
+review REVIEW