Add BuildTestFixture to AbstractIntegrationSpec
Thereby making methods like `multiProjectBuild`
available to all integration tests. By default,
`BuildTestFixture` now creates the build in the
root directory since this is in line with the behaviour of
the other methods in `AbstractIntegrationSpec`.

Moreover, `BuildTestFixture` now takes a `TestDirectoryProvider`
instead of a `TestFile` in order to have it as a field in
`AbstractIntegrationTest` and initializing the `TestDirectoryProvider`
as late as possible.

GRADLE-3534
+review REVIEW-6188