Fix extra property lookup order for grand children
It looks like for grand children of the root
project we were first looking up `ExtraProperties`
in the root project and then in the child project.

Changing the order could have a performance impact,
so this probably needs to be revisited after having a
look at the performance tests.

The order of `snapshotInheritable()` is now the same as
in `updateDelegates`.

GRADLE-3530 #resolve
+review REVIEW