Make Gradle run when TERM is not set
I didn't manage to reproduce the issue
with an integration test since
I could not remove the TERM variable from the environment
of the forked Gradle process. Even starting it via a shell script
having `unset TERM` before the actual Gradle call didn't
make this error surface. It is very easy to reproduce on the
command line.

GRADLE-2634
+review REVIEW