Introduce a separate `BuildIdentifier` for composites
ProjectComponentIdentifier and ProjectComponentSelector now each have
a BuildIdentifier instance to identify the build that contains the
project. You can create a project identifier for the 'current' build
or from an included build.

At this stage, any ProjectComponentIdentifier for the 'current'
build has a 'current' BuildIdentifier, with name = `null`.
This will be improved to be a true build identifier.