Correct sample for using the ear plugin
The previous version of that sample did not actually package
the .war file inside the ear, but the .jar file. This is because
the 'default' configuration contains the .jar. This is not something
we can change easily, because downstream clients will usually need
the .jar to compile against. We will probably need some usage-aware
dependency resolution to tell apart "I want to compile against this"
versus "I want to package this for deployment"