Introduce the `--flamegraphs` flag
When a performance test is executed with the `--flamegraphs` flag, a report directory
will be generated under the test output build directory containing the flame graphs
for the execution of this test.

This implies:

- that honest profiler will be executed as part of the test
- that the environment variables are set to find honest profiler and the flame graph generator

The output zip of the performance test task will contain the flame graphs, making it possible
to fetch those from the CI server.