Change the `flamegraphs` flag so that it needs to be set to `true`.
The rationale for this change is that when we're going to run those from TeamCity,
a parameter has to be set. It cannot be absent, so we need to set it to a recognizable
value. By setting to `true`, we're going to generate flame graphs. Any other value
will result in the traditional behavior, without profiling or flame graph generation.