Use a single delegating task for each included build
Rather than constructing a separate delegating task for each task
dependency in an included build, we now use a single delegating task
for all tasks in an included build. This means that if the task graph
contains multiple task dependencies on an included build, all of these
tasks will be run in a single build execution, at the time the first
task of them is required.

+review REVIEW-6192