Deprecate `ModuleDependency#getConfiguration` in favor of `ModuleDependency#getTargetConfiguration`
We want to be able to differentiate between two cases when dependencies on projects are declared:

1. the user didn't choose any specific configuration
2. the user chose a configuration, and it can be the same as the default one

The previous implementation required that `getConfiguration` returns a non-null value, making it impossible to
make a difference between the following cases:

`compile project(':')`

and

`compile project(path: ':', configuration: 'default')`

The new `getTargetConfiguration` method returns an `Optional<String>`, which will be `absent` if the user
didn't choose anything specific, and `present` if he did.

This allows us to make an explicit configuration selection take precedence over automatic configuration
selection when attributes are defined.