Do not fail whenever a matching configuration is not found in the target project
Whenever a client configuration defines attributes, and that we cannot find a configuration in the target which
matches the attributes, we had the choice between 2 options:

1. fallback to the `default` configuration
2. fail with a reasonable error message

Option 1. has been retained, because it is backwards compatible. In particular, a consumer configuration like `compile` can
declare that it prefers the `api` variant by adding an attribute (say `type: api`). However not all subprojects may be using
a mechanism to produce an API, and therefore not all subprojects will have configurations tagged with `api`. We don't want
to fail in this case, but rather fallback to the `default` configuration. This is basically the same as depending on an
external dependency: such a dependency doesn't have all the metadata we can attach to a configuration (but it may be
reconstructed, as generating an API jar, but that's a subsequent story).