Fail whenever a client configuration defines attributes and more than one target configuration matches
It is possible for a client configuration to define attributes which are matched by multiple target configurations.
In this case, we are now failing with a reasonable error message, telling which configurations match. This
gives the opportunity to the user to specify an explicit configuration in the dependency, instead of relying on
the generic algorithm.