Reset baselines for failing performance tests
There was a small acceptable regression introduced as part of the
implementation of gradle/task-output-cache#25 which has caused many
performance tests to begin failing.

Rather than investigating exactly which of these tests are failing due
to that change, we are just setting the baseline for all failing tests
to the first nightly build which included the change to see which
performance tests may still be failing for other reasons.