Reset the baseline for more performance tests
This baseline is after the changes went in on September 21st to make
parallel builds work faster. This change reduced contention on a single
global lock, but has introduced some overhead in some use cases.

Specifically, the fully up-to-date builds, full builds which don't use
the daemon or parallel workers, and when you are just resolving a ton of
dependencies.  For now, we are considering those use cases as not
important enough to cause us to keep failing the performance tests.