Inject empty interfaces for missing classes in ProjectBuilder
The code that generates these empty interfaces, to ensure compatibility
of plugins compiled with Gradle 2.x in 3.x, is now used to inject these
interfaces into the system class loader during ProjectBuilder
initialisation. In addition, an integration test is added that
demonstrates how to work around the issue in Gradle 3.0 and 3.1
(as this fix will be part of 3.2).

This allows users to write plugin tests with ProjectBuilder covering the
cross-version support scenario. The behaviour of a real Gradle run,
where the injection happens in the class loader hierarchy, can only be
mimicked by this. This is because, in such test scenarios, the
ProjectBuilder runs in a plain Java application with a flat class loader.

+review REVIEW-6246