Use fixed library versions in build init templates
This removes the dynamic version calculation for libraries from
the build init project and replaces it with fixed versions
(latest released) of the libraries.

The code for updating the version file with newer stable releases,
has been moved to the 'gradle promote' build and will be automatically
executed when a new release cycle is started. In case it is needed,
selected versions can still be manually updated during a release cycle.

+review REVIEW-6282