Add performance test without pushing to local cache
The test measures the most overhead a developer needs to pay when they try to pull results from the cache for every task, but miss every time. The test excludes the overhead of pushing to the cache, as developer machines are not expected to push to the cache in this setup.

+review REVIEW-6283